+++ FRONTEND/portal/htdocs/wp-content/plugins/nivo-slider-lite/includes/class-nivo-slider-i18n.php
@@ -1,0 +1,47 @@
+<?php
+
+/**
+ * Define the internationalization functionality
+ *
+ * Loads and defines the internationalization files for this plugin
+ * so that it is ready for translation.
+ *
+ * @link       http://themeisle.com/plugins/nivo-slider/
+ * @since      3.0.0
+ *
+ * @package    nivo-slider
+ * @subpackage nivo-slider/includes
+ */
+
+/**
+ * Define the internationalization functionality.
+ *
+ * Loads and defines the internationalization files for this plugin
+ * so that it is ready for translation.
+ *
+ * @since      3.0.0
+ * @package    nivo-slider
+ * @subpackage nivo-slider/includes
+ * @author     ThemeIsle <friends@themeisle.com>
+ */
+class Nivo_Slider_i18n {
+
+
+	/**
+	 * Load the plugin text domain for translation.
+	 *
+	 * @since    3.0.0
+	 */
+	public function load_plugin_textdomain() {
+
+		load_plugin_textdomain(
+			'nivo-slider',
+			false,
+			dirname( dirname( plugin_basename( __FILE__ ) ) ) . '/languages/'
+		);
+
+	}
+
+
+
+}