+++ FRONTEND/portal/htdocs/wp-content/plugins/nivo-slider-lite/CHANGELOG.md
@@ -1,13 +1,49 @@
 
-### 1.0.6 - 14/02/2017
-**Changes:** 
-- Development
+ ### v2.0.3 - 2017-05-10 
+ **Changes:** 
+ - Small changes to .js for better integration with Shop-Isle-Pro
+- Fixes height calculations.
+ 
+ ### v2.0.2 - 2017-05-06 
+ **Changes:** 
+ - Removed redundant code.
+ ### v2.0.1 - 2017-05-02
 
-### 1.0.5 - 10/01/2017
-**Changes:** 
-- Development
+**Changes:**
 
-### 1.0.5 - 23/11/2016
-**Changes:** 
-- Added themeisle as author
+- Fixed an issue with style loading for other post types.
 
+### v2.0.0 - 2017-04-28  
+
+**Changes:** 
+ 
+- Added more slider themes.
+- Added responsive option.
+- Added more transition effects.
+- Major code refactor.
+ 
+### 3.0.0 - 17/02/2017
+
+**Changes:**
+
+- Refactored code base
+
+### 2.5.1 - 12/12/2016
+
+**Changes:** 
+
+- added backward compat for license
+
+
+### 2.5.0 - 21/11/2016
+
+**Changes:** 
+
+- Added plugin to themeisle system
+
+
+### 2.5.0 - 21/11/2016
+
+**Changes:** 
+
+- Added plugin into themeisle system